Added ModuleFactory
